pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "stock-analyst"
version = "0.1.0"
description = "AI-powered Indian Stock Analyst — Powered by Claude"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
]

[project.scripts]
stock-analyst = "src.main:main"

[tool.setuptools.packages.find]
include = ["src*"]
